+++ CHANGELOG.md
@@ -5,6 +5,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] (2021-04-25)
+
 ### Added
 - method `IAVLTree.isEmpty()`
 - method `IAVLTree.size()`
@@ -18,7 +20,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.0] (2021-04-25, initial release)
 
 ### Added
-
 Initial version featuring the basic BST operations:
 - Insert values
 - Search values
@@ -28,6 +29,7 @@ Initial version featuring the basic BST operations:
 [gh]: https://github.com/JJWesterkamp/avl-bst
 
 
-[Unreleased]: https://github.com/JJWesterkamp/avl-bst/compare/v0.1.1...HEAD
+[Unreleased]: https://github.com/JJWesterkamp/avl-bst/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/JJWesterkamp/avl-bst/compare/v0.1.1...v0.2.0
 [0.1.1]: https://github.com/JJWesterkamp/avl-bst/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/JJWesterkamp/avl-bst/tree/v0.1.0